apps/ws-manager/requirements.txt
websockets==13.1
python-dotenv
tf2-utilities
uvicorn
aiohttp
fastapi
orjson
motor